requirements.txt
bitmart-python-sdk-api==1.0.0
Flask==2.0.1
numpy==1.21.1
pandas==1.3.1
python-binance==1.0.12
requests==2.26.0
TA-Lib==0.4.19